fix water masking of all water images for fs9
imagetool.exe has a bug where it drops the alpha channel of an image if
it is all one color. So, images that were all water (all black) have
their alpha channels dropped. This is a problem for images with no land,
it means there is no water masking at all and the image shows up with no
masking. I solve this by setting one pixel alpha to 255 in these sorts of
images