works better. islands inside others aren't all water now.
TODO: smarter solution is to do blendpool for those who don't have an
island and coasttwo deepwater two for those that do. doing coasttwo and
deepwatertwo for all inland water means places with a lot of inland
lakes, at low resolution, will be overmasked. I guess user can edit the
s functions in the ini, but let's not make them do that