Fix big flaw when checking whether pointOnEdge
Basically, there could be cases (such as in Lake Okeechobee) where simply checking between two sequential points (A and B) in the shared points list would lead to the erroneous conclusion that a third point (C) was on the edge between A and B. But A and B, although sequential in the points list, didn’t describe a true, shared edge. So the conclusion was incorrect. I solve this by getting the shared edges between the two way’s and then only checking if a point is in the shared edges.

Other improvements:

1) sanity check mergePointToPoint by making sure both ways to be merged are lines (not closed ways)
2) fix minor bug in appendParts (breaking if not in shared edge was placed incorrectly)
3) in mergeRivers - use mergeWithWay instead of mergeEdgeToEdge. Given 1) and that inland water bodies I think will always be closed (not sure of this), I don’t think this is necessary. But it’s good to have for sanity reasons