initial work on using pools
it seems this was the solution all along. I fixed the in vs out problem
by:

1) using the same update relations logic you currently had
2) sort the ways by size, starting from the smallest ones. This ensures
that when they are put into the kml, the smallest ones come first. This
is important as the ones that come first (on top) overwrite the ones
that come later (on the bottom). Having created manual polygons in
google earth and figuring out the ordering would have saved me all this
time... I never expected the coast/deepwater method I ended up taking
would have so many edge cases and pitfalls due to the way the OSM data
is