fix large inland waters (masked like the sea) not being masked
there was a floating point compare bug. has_water was being set to 3
instead of 7
despite mesh version being 1.3 because it was comparing a float to a
double... this caused inland sea's (tri_type of 4) to not be masked
because this would trigger the continue, thereby skipping these
tri's...:
(tri_type & has_water) == 0

I guess I am dumb for not just setting has_water to 7. since this code
is from ortho4xp v1.3... but I just blindly copied without thinking
much...