fix scenproc script full path incorrect when resample is running
fset sets the environment dir to the working dir before calling
resample. this means that getfullpath will use the working dir, causing
scenproc scripts to not be found. to circumvent this, we keep them in
Scenproc_scripts folder and just pass the name of the file in that
folder in fsearthtiles.ini