add coords of current single area to areaearthinfo file
this allows water masking to be created for the specified area. water
masking was broken with the new method that was preventing downloading
of excess tiles. this fixes it. basically, in areaearthinfo file, the
fsresampled coords would go in if picture perfect resamplign mode was
set, screwing up with the water masking tilestodownload logic