fix Console.WriteLine crashing on subsequent runs of RunScenproc
set Console stdin and stdout again, or get crashes on subsequent calls of RunScenproc. Why? it appears these handles
are set when the program first starts, even though we don't have a console. Free'ing and the alloc'ing a new console
causes the handles to not be set correctly to the new console, and Console.WriteLine crashes with an invalid handle exception
see: https://stackoverflow.com/questions/42612872/exception-when-using-console-window-in-a-form-application