fix corrupt images
I'm not entirely sure why some images were being corrupt. I think it has to do with the GCHandle,
which is freed elsewhere in the code. cloning the code changed in this commit and not doing it like
I've done in this commit causes some images to be duplicate of others or have black squares (aka be corrupt)