fix preview display not updating when a new lay provider chosen from combo box
due to the lay providers using the previous non lay provider
mService. so it was using the previous non lay cached tile.

The trick lies in the fact that the new lay tiles don't use
mService index. instead, they are identified by a new string
which contains the name of the layservice that they represent.

Basically, this commit fixes the caching of the new lay service tiles