don't create masks and resample in one thread
instead, do the image processing and mask creation in one
multithreaded queue thread, and the image compilation in the separate
multithreaded queue's thread

this way, we aren't waiting for resample to finish before downloading
new areas. instead, we are waiting on the image processing and the
creation of masks (both of the ram heavy tasks) before downloading new
areas (to prevent oom). makes no sense to wait on resample when that
isn't the main ram issue + it's slower if have to wait on
resample before downloading vs waiting on image processing + mask
creation